demand_forecasting_model/requirements.txt
# Core ML
tensorflow==2.14.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.0

# Data
pyarrow==12.0.1
sqlalchemy==2.0.20
google-cloud-bigquery==3.12.0
google-cloud-storage==2.10.0

# Feature Engineering
tsfresh==0.20.0
statsmodels==0.14.0
hyperopt==0.2.7
geohash2==1.1.0

# API & Serving
fastapi==0.104.1
uvicorn==0.24.0
grpcio==1.59.0
protobuf==4.24.1

# Evaluation & Monitoring
prometheus-client==0.18.0
optuna==3.14.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0
jupyter==1.0.0

# Evaluation
plotly==5.17.0
matplotlib==3.8.1
seaborn==0.13.0
